app/requirements.txt
fastapi>=0.68.0
pydantic>=2.0
uvicorn>=0.24.0
python-dotenv
langchain
langchain-openai
jsonschema
